+++ matrix-sdk/src/main/java/org/matrix/androidsdk/call/IMXCall.java
@@ -23,6 +23,9 @@ import com.google.gson.JsonObject;
 import org.matrix.androidsdk.MXSession;
 import org.matrix.androidsdk.data.Room;
 import org.matrix.androidsdk.rest.model.Event;
+import org.webrtc.PeerConnection;
+
+import java.util.List;
 
 import javax.annotation.Nullable;
 
@@ -336,4 +339,6 @@ public interface IMXCall {
      * @return true if video recording is muted, false otherwise
      */
     boolean isVideoRecordingMuted();
+
+    List<PeerConnection.IceServer> getIceServers();
 }