+++ matrix-sdk/src/main/java/org/matrix/androidsdk/call/MXWebRtcCall.java
@@ -209,7 +209,7 @@ public class MXWebRtcCall extends MXCall {
      * @param context    the context
      * @param turnServer the turn server
      */
-    public MXWebRtcCall(MXSession session, Context context, JsonElement turnServer) {
+    public MXWebRtcCall(MXSession session, Context context, JsonElement turnServer, String defaultIceServerUri) {
         if (!isSupported(context)) {
             throw new AssertionError("MXWebRtcCall : not supported with the current android version");
         }
@@ -228,6 +228,16 @@ public class MXWebRtcCall extends MXCall {
         mSession = session;
         mContext = context;
         mTurnServer = turnServer;
+        if (!TextUtils.isEmpty(defaultIceServerUri)) {
+            try {
+                if (!defaultIceServerUri.startsWith("stun:")) {
+                    defaultIceServerUri = "stun:" + defaultIceServerUri;
+                }
+                defaultIceServer = new PeerConnection.IceServer(defaultIceServerUri);
+            } catch (Throwable e) {
+                Log.e(LOG_TAG, "MXWebRtcCall constructor  invalid default stun" + defaultIceServerUri);
+            }
+        }
     }
 
     /**
@@ -566,39 +576,10 @@ public class MXWebRtcCall extends MXCall {
         }
 
         // build ICE servers list
-        List<PeerConnection.IceServer> iceServers = new ArrayList<>();
-
-        if (null != mTurnServer) {
-            try {
-                String username = null;
-                String password = null;
-                JsonObject object = mTurnServer.getAsJsonObject();
-
-                if (object.has("username")) {
-                    username = object.get("username").getAsString();
-                }
-
-                if (object.has("password")) {
-                    password = object.get("password").getAsString();
-                }
-
-                JsonArray uris = object.get("uris").getAsJsonArray();
-
-                for (int index = 0; index < uris.size(); index++) {
-                    String url = uris.get(index).getAsString();
-
-                    if ((null != username) && (null != password)) {
-                        iceServers.add(new PeerConnection.IceServer(url, username, password));
-                    } else {
-                        iceServers.add(new PeerConnection.IceServer(url));
-                    }
-                }
-            } catch (Exception e) {
-                Log.e(LOG_TAG, "## createLocalStream(): Exception in ICE servers list Msg=" + e.getMessage(), e);
-            }
+        List<PeerConnection.IceServer> iceServers = getIceServers();
+        if (iceServers.isEmpty() && defaultIceServer != null) {
+            iceServers.add(defaultIceServer);
         }
-
-        Log.d(LOG_TAG, "## createLocalStream(): " + iceServers.size() + " known ice servers");
 
         // define at least on server
         if (iceServers.isEmpty()) {